+++ lib/src/selectors/color_well.dart
@@ -114,8 +114,12 @@ class _MacosColorWellState extends State<MacosColorWell> {
           await _methodChannel.invokeMethod('color_panel', {
             'mode': '${widget.defaultMode}',
           });
+        } on MissingPluginException {
+          throw MissingPluginException(
+            'Cannot open NSColorPanel on platforms other than macOS.',
+          );
         } catch (e) {
-          debugPrint('$e');
+          debugPrint('Failed to open color panel: $e');
         }
       },
       child: SizedBox(