+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [1.1.0]
+* New functionality for `MacosSearchField`
+  * Shows a list of search results in an overlay below the field
+  * A result can be selected and customized.
+* A `MacosOverlayFilter` widget can now be used to apply the blurry "frosted glass" effect on surfaces.
+* New widget: `CustomToolbarItem` that enables any widget to be used in the `Toolbar`.
+
+## [1.0.1]
+* Improvements to the graphical `MacosTimePicker`
+  * Better color gradient on the border
+  * Better inner shadow
+  * Minor size adjustments
+  * API improvements
+* Throw an exception if `MacosColorWell` is clicked on a non-macOS platform
+
 ## [1.0.0+1]
 * Minor documentation fix for [MacosColorWell]